fs: Move more data structures from fs.h to fs-internals.h
In order to keep the interface between kernel subsystems as small and
clean as possible most data structures where moved from the public to
the internal header. The only structure to be keept is the task_fs_data
since it's part of the inter-subsystem api.
